% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assist_net.R
\name{assist_net}
\alias{assist_net}
\title{Assist Network}
\usage{
assist_net(.data)
}
\arguments{
\item{.data}{play-by-play data frame returned from w_get_pbp_game function}

\item{team}{Team to create network for. Can be bare home or away variable from play-by-play object or quoted team name}

\item{node_col}{Color of nodes in network. Can be selected but defaults 
to primary team color and then a backup if one is not found.}

\item{three_weights}{Logical indicating whether to give extra weight for assisted three point shots.
If TRUE, assisted three-point shots will be given weight 1.5 (as opposed to weight 1). Default = `TRUE`.}

\item{threshold}{Number between 0-1 indicating minimum percentage of team assists/baskets a player needs to exceed to be included in network. Default = 0.}

\item{message}{User supplied plot title to overwrite default plot title, if desired. Default = `NA`.}

\item{listing}{Allows disabling the return of list of network statistics}
}
\value{
List with network statistics
\itemize{
 \item{"clust_coeff"} - Network Clustering Coefficient
 \item{"page_ranks"} - Player Page Ranks in network
 \item{"hub_scores"} - Player Hub Scores in network
 \item{"auth_scores"} - Player Authority Scores in network
 \item{"ast_freq"} - Player percentage of team's assists
 \item{"shot_freq"} - Player percenatge of scoring on team's assisted baskets
 }
 @importFrom dplyr group_by
 @importFrom dplyr summarize
}
\description{
This function produces an assist network visualization for a team for a single game (or collection of games).
}
