% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_score_diff.R
\name{average_score_diff}
\alias{average_score_diff}
\title{Average Score Differential}
\usage{
average_score_diff(pbp_data)
}
\arguments{
\item{pbp_data}{Play-by-play data returned from w_get_pbp_game}
}
\value{
Average score differential
}
\description{
This function returns a single value for the average score differential.
}
\details{
Computes Time-Based Average Score Differential for the Home Team
}
