% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_win_prob.R
\name{average_win_prob}
\alias{average_win_prob}
\title{Average Win Probability}
\usage{
game_flow(pbp_data, home_col, away_col)
}
\arguments{
\item{pbp_data}{Play-by-play data returned from w_get_pbp_game}
}
\value{
Average Win Probability
}
\description{
This function returns the average win probability for the home team.
}
