% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_flow.R
\name{game_flow}
\alias{game_flow}
\title{Game Flow Chart}
\usage{
game_flow(.data, home_col, away_col)
}
\arguments{
\item{.data}{Play-by-play data returned from w_get_pbp_game}

\item{home_col}{Color of home team for chart. Can be selected but defaults 
to primary team color.}

\item{away_col}{Color of away team for chart. Can be selected but defaults 
to primary team color.}
}
\value{
A ggplot2 object
}
\description{
Creates a game flow chart with ggplot2.
}
\details{
This function takes play-by-play data returned from w_get_pbp_game
and returns a ggplot2 object to show the game flow.
}
\examples{
pbp_data <- w_get_pbp_game("401176897")
game_flow(pbp_data, "red", "black")
}
