% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{get_line}
\alias{get_line}
\title{Get Line}
\usage{
get_line(.data)
}
\arguments{
\item{.data}{play-by-play data frame returned from w_get_pbp_game function}
}
\value{
Line
}
\description{
Gets the line (in points) of the game from WHERE?
}
