% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shot_charts.R
\name{get_shot_locs}
\alias{get_shot_locs}
\title{Get Shot Locations}
\usage{
get_shot_locs(pbp_data, home_col, away_col)
}
\arguments{
\item{game_ids}{Vector of ESPN game_ids}
}
\value{
A data-frame with shot details (including location) for given game_ids
}
\description{
Gets and plots shot locations for a game.
}
