% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ids}
\alias{ids}
\title{Team Name Data}
\format{
A data frame with ESPN team names to be used in the ncaahoopR package.
\describe{
 \item{team}{the name of the team to be suplied to function in ncaahoopR package}
 \item{id}{team id; used in ESPN URLs}
 \item{link}{team URL link; used in ESPN URLs}
}
}
\usage{
data(ids)
}
\description{
Provides a data frame with ESPN team names for use in the wncaahoopR package.
}
\details{
Simply look for the spelling of your deired team in the team column
to use in ncaahoopR functions.
}
\examples{
data(ids)

}
\keyword{datasets}
