% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\docType{data}
\name{pbp_data}
\alias{pbp_data}
\title{Get Line}
\format{An object of class \code{rlang_fake_data_pronoun} of length .}
\usage{
get_line(pbp_data)
}
\arguments{
\item{.data}{play-by-play data frame returned from w_get_pbp_game function}
}
\value{
Line
}
\description{
Gets the line (in points) of the game from WHERE?
}
\keyword{datasets}
