% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secs_to_model.R
\name{secs_to_model}
\alias{secs_to_model}
\title{Seconds to Model Helper}
\usage{
secs_to_model(sec, msec)
}
\value{
list of minutes and seconds remaining
}
\description{
Get appropriate model for time remaining.
}
\keyword{internal}
