% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_get_master_schedule.R
\name{w_get_master_schedule}
\alias{w_get_master_schedule}
\title{Get Master Schedule}
\usage{
w_get_master_schedule(date)
}
\arguments{
\item{date}{Date for which to get schedule (YYYY-MM-DD)}
}
\value{
A data-frame of the day's schedule of games
}
\description{
Scrapes ESPN Play-by-Play data for all games on a given date.
}
\details{
Gets schedule for all games on a given date
}
