% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wcourt}
\alias{wcourt}
\title{A data frame with women's NCAA court dimensions.}
\format{
A data frame with women's NCAA court dimensions
}
\usage{
wcourt
}
\description{
A data frame with women's NCAA court dimensions.
}
\details{
use this dataframe with geom_polygon to draw a NCAA court.
}
\examples{
data(wcourt)

}
\keyword{datasets}
