% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_chart.R
\name{wp_chart}
\alias{wp_chart}
\title{Win Probability Chart}
\usage{
wp_chart(pbp_data, home_col = NULL, away_col = NULL, show_legend = TRUE)
}
\arguments{
\item{pbp_data}{Play-by-play data returned from w_get_pbp_game}

\item{home_col}{Color of home team for chart}

\item{away_col}{Color of away team for chart}

\item{show_legend}{Logical indicating whether or not to display legend and min win probability
on the chart. Default = TRUE}
}
\description{
Renders win proability chart for desired game
}
