% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wp_compute.R
\name{wp_compute}
\alias{wp_compute}
\title{Win Probability Function}
\usage{
wp_compute(x)
}
\value{
A single number, a probability between 0 and 1.
}
\description{
Computes the win probability, given a game state.
}
\keyword{internal}
