% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wp_hoops}
\alias{wp_hoops}
\title{A data frame containing}
\format{
A data frame with win probabilities
}
\usage{
wp_hoops
}
\description{
A data frame containing
}
\details{
use this data frame for various win probability calculations
}
\examples{
data(wp_hoops)

}
\keyword{datasets}
