% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_score_diff.R
\name{average_score_diff}
\alias{average_score_diff}
\title{Average Score Differential}
\usage{
average_score_diff(pbp_data)
}
\arguments{
\item{.data}{play-by-play data frame returned from w_get_pbp_game function}
}
\value{
Average score differential
}
\description{
Returns a single value for the average score differential from home team's perspective.
}
