% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_win_prob.R
\name{average_win_prob}
\alias{average_win_prob}
\title{Average Win Probability}
\usage{
average_win_prob(.data)
}
\arguments{
\item{.data}{play-by-play data frame returned from w_get_pbp_game function}
}
\value{
Average Win Probability
}
\description{
This function returns the average win probability for the home team.
}
