% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_excitement_index.R
\name{game_excitement_index}
\alias{game_excitement_index}
\title{Game Excitement Index}
\usage{
game_excitement_index(.data)
}
\arguments{
\item{.data}{play-by-play data frame returned from w_get_pbp_game function}
}
\value{
GEI--Game Exictement Index
}
\description{
Computes Game Excitement Index for Desired Game
}
