% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_line.R
\name{get_line}
\alias{get_line}
\title{Get Line}
\usage{
get_line(pbp_data)
}
\arguments{
\item{pbp_data}{Play-by-play data returned from w_get_pbp_game}
}
\value{
Line
}
\description{
Gets the line (in points) of the game from WHERE?
}
