% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_get_pbp_game.R
\name{w_get_pbp_game}
\alias{w_get_pbp_game}
\title{Get Game Play-by-Play Data}
\usage{
w_get_pbp_game(game_ids)
}
\arguments{
\item{game_ids}{Vector of ESPN game-IDs}
}
\value{
A data-frame of the Play-by-Play data fror desired games.
}
\description{
Scrapes ESPN Play-by-Play data for the desired games.
}
