% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_ncaa_pbp_scraper.R
\name{w_get_roster}
\alias{w_get_roster}
\title{Get Team Roster}
\usage{
w_get_roster(team)
}
\arguments{
\item{team}{Team to get roster for}
}
\value{
A data-frame of the team's roster for current season
}
\description{
Gets team roster for current season.
}
