% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_get_roster.R
\name{w_get_roster}
\alias{w_get_roster}
\title{Get Team Roster}
\usage{
get_roster(team)
}
\arguments{
\item{team}{Team to get roster for}
}
\value{
A data-frame of the team's roster for current season. Largely internal.

A data-frame of the Play-by-Play data fror desired games.
}
\description{
Gets team roster for current season.
}
