% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freshScraperDraft.R, R/w_ncaa_pbp_scraper.R
\name{w_get_schedule}
\alias{w_get_schedule}
\title{Get Team Schedule}
\usage{
w_get_schedule(team)

w_get_schedule(team)
}
\arguments{
\item{team}{Team to get schedule for}

\item{team}{Team to get schedule for}
}
\value{
A data-frame of the team's schedule for current season

A data-frame of the team's schedule for current season
}
\description{
Gets team schedule for current season.

Gets team schedule for current season.
}
